+++ CHANGES.rst
@@ -4,6 +4,7 @@
 Fixed
 ^^^^^
 - The SCIM endpoint correctly manages patching user passwords.
+- The SQL backend can perform exact match searches on JSON list attributes. :issue:`278`
 
 [0.0.82] - 2025-08-26
 ---------------------